cloud/requirements.txt
flask
flask-restx
flask-SQLAlchemy
flask-script
flask-migrate
flask_testing

sqlalchem==1.3.23 #1.4.0 breaks something right now

sentry-sdk[flask]
blinker